(Refact) seperate wish company list component from Mypage.tsx
